refactor: Update imports + usage for LabeledSlider
