+++ src/components/LabeledSlider.tsx
